Update to README.md
I updated it with my name and email on the first line. Secondly, I included a brief description of my GitHub account on the second line. Thirdly, I incorporated a brief description of the repository I created on the third line. Lastly, a brief description of the steps I took to clone the repository onto my local machine on the fourth line.